requirements.txt
streamlit
chromadb
groq
PyPDF2
python-dotenv
onnxruntime
transformers